Use the API service to pull back projects rather than fetch directly
