+++ app/src/Config/index.example.ts
@@ -1,4 +1,9 @@
 export const Config = {
-  API_URL: 'https://jsonplaceholder.typicode.com/',
-  BASE_URL: 'http://localhost:5000',
+  // This will eventually be what the STA_BASE_URL is. The API_URL
+  // it is set to an empty string so we can use both for now.
+  API_URL: '',
+  // Example API Usage
+  EXAMPLE_USER_URL: 'https://jsonplaceholder.typicode.com/',
+  // STA Base API URL
+  STA_BASE_URL: 'http://localhost:5000',
 }
