+++ app/src/Services/modules/index.ts
@@ -1,1 +1,2 @@
+export { projectsApi } from './projects'
 export { userApi } from './users'
