+++ app/src/Services/modules/projects/index.ts
@@ -1,0 +1,24 @@
+import { api } from '../../api'
+import fetchAll from './fetchAll'
+
+export const projectsApi = api.injectEndpoints({
+  endpoints: build => ({
+    fetchAll: fetchAll(build),
+  }),
+  overrideExisting: false,
+})
+
+export const { useLazyFetchAllQuery } = projectsApi
+
+export interface Project {
+  name: string
+  client: string
+  role: string
+  description: string
+  skills: string[]
+  hours: number
+  required: boolean
+  buddying: boolean
+}
+
+export type Projects = Project[]
